Extract AbstractModel::fromArray|toArray in Subclass for Embedded (custom) objects in Doctrine (eg Embeddables).
